requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
sentence-transformers==2.7.0
numpy==2.1.3
pydantic==2.11.7
python-dotenv==1.1.1
chromadb==0.5.20
langchain-text-splitters==1.1.0
aiosqlite==0.20.0
sqlalchemy==2.0.31
pypdf==4.2.0
python-docx==1.1.2
qdrant-client==1.16.2

